fix: re-organize homepage components
