+++ src/components/Homepage/TicketItem.jsx
