+++ src/components/Homepage/__test__/CommunicateThroughSection.test.js
