+++ src/components/Homepage/__test__/TicketItem.test.js
