+++ src/components/Homepage/__test__/TicketSection.test.js
