+++ src/components/__test__/HomePage.test.js
@@ -1,0 +1,8 @@
+import renderer from "react-test-renderer";
+
+import HomePage from "../HomePage";
+
+it("renders correctly", () => {
+    const tree = renderer.create(<HomePage />).toJSON();
+    expect(tree).toMatchSnapshot();
+});
